--- fluffy_poltergeist/entities/entities/prop_spawner/shared.lua
@@ -1,2 +1,0 @@
-
-ENT.Type = "point"
