+++ fluffy_poltergeist/gamemode/cl_init.lua
@@ -1,5 +1,1 @@
 include('shared.lua')
-
-function GM:HUDDrawTargetID()
-	return false
-end
