--- fluffy_poltergeist2/fluffy_poltergeist2.txt
@@ -1,6 +1,0 @@
-"minigames"
-{
-	"base"	"fluffy_mg_base"
-	"title"	"Poltergeist"
-    "maps"	"^pg_"
-}
