--- fluffy_poltergeist2/gamemode/cl_init.lua
@@ -1,1 +1,0 @@
-include('shared.lua')
